+++ ISO8211.xcodeproj/xcuserdata/christopheralford.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -14,8 +14,8 @@
             filePath = "ISO8211/CatalogProvider.swift"
             startingColumnNumber = "9223372036854775807"
             endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "119"
-            endingLineNumber = "119"
+            startingLineNumber = "120"
+            endingLineNumber = "120"
             landmarkName = "viewSubfield(poSFDefn:pachFieldData:nBytesRemaining:)"
             landmarkType = "7">
             <Locations>
@@ -55,54 +55,6 @@
       <BreakpointProxy
          BreakpointExtensionID = "Xcode.Breakpoint.FileBreakpoint">
          <BreakpointContent
-            uuid = "EFB2AAA6-B1A3-4C57-9ABF-29DAF134CB44"
-            shouldBeEnabled = "Yes"
-            ignoreCount = "0"
-            continueAfterRunningActions = "No"
-            filePath = "ISO8211/ISO8211Lib/DDFSubfieldDefinition.swift"
-            startingColumnNumber = "9223372036854775807"
-            endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "97"
-            endingLineNumber = "97"
-            landmarkName = "setFormat(_:)"
-            landmarkType = "7">
-            <Locations>
-               <Location
-                  uuid = "EFB2AAA6-B1A3-4C57-9ABF-29DAF134CB44 - 5aa697f464852141"
-                  shouldBeEnabled = "Yes"
-                  ignoreCount = "0"
-                  continueAfterRunningActions = "No"
-                  symbolName = "ISO8211.DDFSubfieldDefinition.setFormat(Swift.String) -&gt; Swift.Bool"
-                  moduleName = "ISO8211"
-                  usesParentBreakpointCondition = "Yes"
-                  urlString = "file:///Users/christopheralford/Developer/ISO8211/ISO8211/ISO8211Lib/DDFSubfieldDefinition.swift"
-                  startingColumnNumber = "9223372036854775807"
-                  endingColumnNumber = "9223372036854775807"
-                  startingLineNumber = "98"
-                  endingLineNumber = "98"
-                  offsetFromSymbolStart = "3503">
-               </Location>
-               <Location
-                  uuid = "EFB2AAA6-B1A3-4C57-9ABF-29DAF134CB44 - 5aa697f464852122"
-                  shouldBeEnabled = "Yes"
-                  ignoreCount = "0"
-                  continueAfterRunningActions = "No"
-                  symbolName = "ISO8211.DDFSubfieldDefinition.setFormat(Swift.String) -&gt; Swift.Bool"
-                  moduleName = "ISO8211"
-                  usesParentBreakpointCondition = "Yes"
-                  urlString = "file:///Users/christopheralford/Developer/ISO8211/ISO8211/ISO8211Lib/DDFSubfieldDefinition.swift"
-                  startingColumnNumber = "9223372036854775807"
-                  endingColumnNumber = "9223372036854775807"
-                  startingLineNumber = "97"
-                  endingLineNumber = "97"
-                  offsetFromSymbolStart = "2863">
-               </Location>
-            </Locations>
-         </BreakpointContent>
-      </BreakpointProxy>
-      <BreakpointProxy
-         BreakpointExtensionID = "Xcode.Breakpoint.FileBreakpoint">
-         <BreakpointContent
             uuid = "2A26E9C1-97D6-4260-AA4D-F57705974D2E"
             shouldBeEnabled = "Yes"
             ignoreCount = "0"
@@ -114,54 +66,6 @@
             endingLineNumber = "91"
             landmarkName = "getRepeatCount()"
             landmarkType = "7">
-         </BreakpointContent>
-      </BreakpointProxy>
-      <BreakpointProxy
-         BreakpointExtensionID = "Xcode.Breakpoint.FileBreakpoint">
-         <BreakpointContent
-            uuid = "4BAAB930-5894-4359-883F-AFC09A6628AC"
-            shouldBeEnabled = "Yes"
-            ignoreCount = "0"
-            continueAfterRunningActions = "No"
-            filePath = "ISO8211/ISO8211Lib/DDFSubfieldDefinition.swift"
-            startingColumnNumber = "9223372036854775807"
-            endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "417"
-            endingLineNumber = "417"
-            landmarkName = "getDataLength(pachSourceData:nMaxBytes:pnConsumedBytes:)"
-            landmarkType = "7">
-            <Locations>
-               <Location
-                  uuid = "4BAAB930-5894-4359-883F-AFC09A6628AC - d2a6138985bb5cfc"
-                  shouldBeEnabled = "Yes"
-                  ignoreCount = "0"
-                  continueAfterRunningActions = "No"
-                  symbolName = "ISO8211.DDFSubfieldDefinition.getDataLength(pachSourceData: Foundation.Data, nMaxBytes: Swift.Int, pnConsumedBytes: inout Swift.Optional&lt;Swift.Int&gt;) -&gt; Swift.Int"
-                  moduleName = "ISO8211"
-                  usesParentBreakpointCondition = "Yes"
-                  urlString = "file:///Users/christopheralford/Developer/ISO8211/ISO8211/ISO8211Lib/DDFSubfieldDefinition.swift"
-                  startingColumnNumber = "9223372036854775807"
-                  endingColumnNumber = "9223372036854775807"
-                  startingLineNumber = "418"
-                  endingLineNumber = "418"
-                  offsetFromSymbolStart = "232">
-               </Location>
-               <Location
-                  uuid = "4BAAB930-5894-4359-883F-AFC09A6628AC - d2a6138985bb5c1f"
-                  shouldBeEnabled = "Yes"
-                  ignoreCount = "0"
-                  continueAfterRunningActions = "No"
-                  symbolName = "ISO8211.DDFSubfieldDefinition.getDataLength(pachSourceData: Foundation.Data, nMaxBytes: Swift.Int, pnConsumedBytes: inout Swift.Optional&lt;Swift.Int&gt;) -&gt; Swift.Int"
-                  moduleName = "ISO8211"
-                  usesParentBreakpointCondition = "Yes"
-                  urlString = "file:///Users/christopheralford/Developer/ISO8211/ISO8211/ISO8211Lib/DDFSubfieldDefinition.swift"
-                  startingColumnNumber = "9223372036854775807"
-                  endingColumnNumber = "9223372036854775807"
-                  startingLineNumber = "417"
-                  endingLineNumber = "417"
-                  offsetFromSymbolStart = "229">
-               </Location>
-            </Locations>
          </BreakpointContent>
       </BreakpointProxy>
    </Breakpoints>
